gnome: Remove old aliases
Had to move the aliases outside of the scope or packages in the scope
would have the throws injected instead of the proper packages
from the parent scope.

Also removed the gnome3 alias since that was mostly for internal use
– I doubt people would use gnome3.gnome3.package attribute paths.
It does not really work when moved out of the scope
and it is no longer used in the scope anyway.