nixos/tests: remove broken prosody-mysql test
The test has been broken for some time and the test errors are
non-obvious. None of the current maintainers know how to fix it so it is
better to get rid of it then to keep a continously failing test.

(cherry picked from commit 4369bebd9a32658ded22b580886587cdc577a29d)