google-chrome: passthrough CHROME_WRAPPER environment variable
The XDG desktop menu entries created by Chrome for Progressive Web Apps
are currently broken due to Exec being set to the path of the upstream
Chrome wrapper instead of the Nixpkgs wrapper. This causes Chrome to
crash becaues LD_LIBRARY_PATH is not set.

Chrome obtains the path to be included in the menu entry from the
CHROME_WRAPPER environment variable, which is currently set by the
upstream wrapper to its own path. By setting the variable to the path of
the Nixpkgs wrapper instead, launching PWAs works as expected.

https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/shell_integration_linux.cc
https://source.chromium.org/chromium/chromium/src/+/main:chrome/installer/linux/common/wrapper