nodePackages.typescript-language-server: typescript dependecy as fallback
Make possible to use a different typescript version.
If there is already a typescript binary on your PATH, probably you want
to use that. This way it's possible to use a different typescript
version (with nix shell, direnv, npm, ...), but still have a fallback
version, so the lsp server doesn't fail to start

Related to #73119