pathDerivation: Copy path
Otherwise you end up with a derivation that refers to the original
path (which is not in the Nix store and not accessible to
builders). This caused the derivation paths for the docbookrx package
(removed on master) to depend on the location of the nixpkgs source
tree.

(cherry picked from commit 55ae086747caa71b92f6713a58a909acef944705)