nixos/thelounge: disable integrated package management when appropriate
When using The Lounge's NixOS module to manage packages, the packages are located in the Nix store.

Commands such as `thelounge install` need to write to the directory containing the packages. As such, these commands are disabled.
Additionally, The Lounge's method of checking for outdated packages involves calling Yarn, which requires a lockfile to exist in the packages directory. This module does not generate a lockfile, so this functionality is also disabled.