sqlite-interactive: drop in favor of enabling readline support by default
When using sqlite3 repl without readline support backspace does not work in certain situations.
Also arrow keys to navigate and move in history do not work and print the escape sequences.
It is not obvious why this happens and that you need to use sqlite-interactive instead.
So dropping the option all together to make the repl easier to use.