k3s: v1.22.3+k3s1 -> 1.23.3+k3s1
This PR also reworks the packaging of k3s significantly.

Notably, it makes the following changes:

1. buildGoModule instead of buildGoPackage + mkDerivation (for the final one)
2. Merges the final two derivations (k3s user binary + wrapper script stuff) into one
3. Reduces the amount of patching & usage of upstream packaging scripts

And of course updates to 1.23

I'm not really sure yet if this is significantly better than the
previous packaging model, but it at least seems a little nicer I guess?
buildGoModule is nice, even if we have to hack around it quite a bit to
get a 'go generate' in there.