munt: remove
Munt is a monorepo, and it uses multiple directories for their various
subprojects. While there is an attempt to unified release tags for future
releases, for now we will rely on this current state of things and release each
useful program "independently".

This specific patch is just for cleanup purposes.