fastlane: 2.171.1 -> 2.204.2
This originally was an upgrade to 2.204.1 to workaround an App Store
Connect regression but Apple has since fixed their API. So now this is
just a "wouldn't it be nice to update this package" commit.

Original:
App Store Connect has a bug that prevents build uploads. This new
version of fastlane has a workaround.