qtwebkit: use python38 to build instead of python2
python38 appears to work just as well, so it seems better to use that rather
than python2.  This also resolves some build flakiness seen due to parallel
invocations of python2 that are fixed in python3.

Note that the scripts aren't compatible with python39 or later, some patching
would be required to resolve that.