spruce: Fix build to use go_1_17 and teach go-module a trick.
The root cause of the bug here is that while the go.mod file was updated
to say 1.17, it was not actually run through go.mod 1.17, which ensures
that all deps are enumerated in the go.mod and require files. It
correctly fixes this when run - updating go.mod and go.sum

Note that since these deps are recursive child deps of the existing deps, this
hash should be stable.