irssi: switch from tarball to git tag
I've just switched irssi to build from the 1.2.3 tag in the git tree, as
the build was failing on apple silicon when linking libirc_proxy.la in
dir `src/irc/proxy`

I kinda just switched it and it built. I did not dig very deep on the
reasons behind why this fixed things. I am hoping that Linux still
builds, as i have not tested it. Some notes:

- The sys_lib_search_path done by fixLibtool in the configure phase now
  no longer seems to update with the NIX_LDFLAGS. It appears that this
  is because the newer libtool script does not start the line with
  'eval', so the sed does not match.
- My system still builds, despite the previous item. I noticed that on
  this new build, my system defines `allow_undefined_flag` in the
  libtool script to something nonempty. I guess that may be the reason
  everything still builds and works at runtime.