phinger-cursors: fix installPhase
Running `cp -r ./ $out/share/icons` to copy the icon files to
`$out/share/icons` ends up copying the `env-vars` file as well (since
sourceRoot = .), which can lead to collisions when building the system
path.

This uses a wildcard instead, which relies on the assumption that all
variants start with `phinger-cursors`.