bind: 9.16.25 -> 9.18.0
https://downloads.isc.org/isc/bind9/9.18.0/doc/arm/html/notes.html#notes-for-bind-9-18-0

The new dependency nghttp2 is required for DNS-over-HTTPS support.
Using jemalloc as the memory allocator is now recommended according to the release notes.
Without the change to the configureFlags, configure would complain:
configure: error: --with-gssapi requires yes|no|auto|/path/to/krb5-config
The patch also needed adjusting as it didn't apply anymore.