terraform: fix overrideAttrs with passthru attributes
If plugins are specified in a call to `terraform.withPlugins`, the
standard override functions do not act on the wrapper derivation that
a user sees but on the underlying Terraform derivation.

Due to this, overriding attributes does nothing as they are not passed
through to the "outer" wrapper derivation.

This change fixes this for at least passthru attributes, but other
applications of `.overrideAttrs` (and potentially of the other
overrides) will do nothing (except compute a bit internally).

This fixes #158620 at least partially.