kubernetes: 1.22.6 -> 1.23.3
Note that 1.23 will be the last release to support docker-shim; which is what the kubernetes NixOS module
is still currently using.

For the bump to 1.24 we will need to move to a CRI runtime like CRI-O or containerd. Both are
packaged in nixpkgs already.

The CRI-O maintainers are nix users, so I have a slight preference on picking CRI-O in the
hope to improve the stability. But they should be interchangeable given they both communicate
with kubernetes through the CRI GRPC API.