hydra: remove multi-version support
This is a legacy of when a "stable" version of hydra existed, and this
probably won't be coming back.

This also removes some obsolete bits which have no effect on the
result of evaluation.

Unlike the original approach in #157072, we're moving the common
expression to "unstable.nix" instead. This allow putting everything in
one commit, satisfies git's rename detection, and remains bisectable.

Co-Authored-By: Bernardo Meurer <bernardo@meurer.org>