haskellPackages.utc: mark as broken
utc has been unmaintained for a while now (last commit was 2016). We
duct taped it to keep it compiling, but now System.Clock is no longer
importable safely. I think it's time to lay it to rest, as it also has
some severe bugs causing it to produce incorrect results.

It's a bit sad, as I think it would be the better time library. Maybe
I'll try to become the hackage maintainer at some point [1].

[1]: https://github.com/lpeterse/haskell-utc/issues/8#issuecomment-786653135