tlaps: 1.4.3 -> 1.4.5
This updates tlaps to the most recent stable version and fixes the
template for the url from which the sources are fetched. Previously, the
tarball was fetched from the "current" remote directory from which old
versions are eventually removed. The package was therefore broken.
It also adds myself (florentc) as a maintainer.