qgroundcontrol: 4.1.4 -> 4.2.0
* qgroundcontrol: 4.1.4 -> 4.2.0 (#159291)

QGroundControl now tries to download a prebuilt x86_64 .deb package for
the Airmap SDK, so we have to disable this feature. Previously, it was
disabled by default. I guess they were trying to make things easier for
users, but it seems pretty half-baked.

* qgroundcontrol: update homepage

The domain seems to have changed from .org to .com.