haskellPackages: fix evaluation with stackage nightly 2022-02-17
I've tried to be conservative and only clean up things that broke eval
which in most cases had to do with uses of versioned packages that have
become the default in Stackage Nightly now. There's probably still
potential for cleaning up some additional workarounds and likely there
are still a lot of new workarounds to be found once we know about the
build failures.

Also note that clean eval only means that nix-env doesn't fail.

Notes:

* HLS now lacks the stylish-haskell and britanny plugins by default,
  since they don't support GHC 9.0.2. We'll need to re-enable them
  for GHC < 9.0.2
* A lot of uses of hspec 2.9.* can probably be removed, as 2.8.* which
  is used in Stackage Nightly is new enough for most packages.