installation-cd-graphical-gnome: add useful shortcuts to the dash
The live image is primarily used for installation so we should make
link to manual as well as other useful tools front and center,
instead of having them buried in the app drawer.

The default GNOME apps can still be found there when the ISO
is used for demonstration purposes.