mips64: have mips64el chips auto-detect abi=n64
This PR adds an additional two-line patch to the code which calculates
a `hostPlatform` based on the `builtins.system` provided by the `nix`
interpreter, which is in turn derived from the runtime CPU.

If the runtime CPU is a 64-bit mips chip running on a little-endian
kernel, and the user has not specified otherwise, this commit will
cause nixpkgs to assume the "n64" ABI.

All mips64 chips are capable of supporting all the mips ABIs.

The n64 ABI was chosen because boost's user-scheduled threading (which
nix requires) currently does not support the n32 ABI -- it lacks the
assembler file which explains how to save and restore the processor
state for that ABI.  This is also the ABI that Debian has chosen for
this platform.