kmod-blacklist-ubuntu: don't refer to grep/xargs
64b4af52961 ("kmod-blacklist-ubuntu: 22-1.1ubuntu1 -> 28-1ubuntu4")
doubled the size of the default initramfs.  This happened because the
upgrade introduced this configuration:

	remove iwlwifi \
	(/sbin/lsmod | grep -o -e ^iwlmvm -e ^iwldvm -e ^iwlwifi | xargs /sbin/rmmod) \
	&& /sbin/modprobe -r mac80211

This meant that the grep and xargs substitutions, which had been
inactive for years, suddenly became active again and became part of
kmod-blacklist-ubuntu's closure.

Since we're already using /run/booted-system for the kmod binaries,
I think it's okay to use it for grep and xargs as well.  Both are
required NixOS packages, so they're guaranteed to be there.

Large increases in initramfs size are problematic, because it's often
not possible for users to do anything about them.  It's not always
possible to increase the size of /boot, because some filesystems like
ZFS don't support being shrunk to make way for a bigger /boot.