linux-rt: Fix incorrect sorting of kernel version numbers when determining latest version
When determining the latest linux-rt kernel version to update to, the list of filenames that `update-rt.sh` fetches from upstream is sorted lexicographically, while `update-rt.sh` assumes it's sorted in version order.  As a result, for example, 5.10.78 is considered newer than 5.10.100.  This change applies `sort --version-sort` to the extracted version numbers to properly identify the latest version.