pinta: re-added missing runtime icons
On PR #155224, we switched to `dotnet build` to copy the Pinta assets to
the derivation. This worked mostly fine, however it seems to have missed
the runtime icons. We didn't catch the issue on #155224 neither,
probably because of some resource cache somewhere.

This commit re-adds the missing runtime icons by copying them manually.