schildichat: 1.9.8-sc.1 -> 1.10.3-sc.0.test.1
The latest release is being held back because there is an issue with
Electron 15 crashing on some platforms.
NixOS seems to be not affected by this issue, and the upstream
developers say this test release should be fine otherwise.
This also means we can have a SchildiChat build that is not using the
EOL/insecure Electron 13.x.